Use hidden field to store selected Crownstone
Before this commit, the selected item in the list was not automatically filled in the input field because the list was not yet created. This was solved by requesting the selected Crownstone together with the Crownstones to create the list. This commit solves this. A hidden field is used to store the selected Crownstone. When the list is created, the Crownstone is selected based on the hidden input field. This is more efficient and saves the value without needing to deploy the code.