Localize user: added list of spheres
A list of spheres is added. The sphere is needed to request users.

This node can be changed to list all users from all spheres. When that is made, the user does not need to select the sphere before selecting the user.
(Idea: Show the sphere name when hovering over the names.)