SSE client get token after re-authentication #6
The SSE client will get the current access token when starting the SSE client. The SSE client node can be connected to the authentication node to get triggered when the user re-authenticates.

Possible problem is that the SSE client starts when the user is re-authenticated, even if the user stopped the SSE client before on purpose.